+++ src/Symfony/Component/Messenger/Tests/Transport/Serialization/PhpSerializerTest.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Messenger\Tests\Transport\Serialization;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Messenger\Envelope;
+use Symfony\Component\Messenger\Exception\MessageDecodingFailedException;
 use Symfony\Component\Messenger\Tests\Fixtures\DummyMessage;
 use Symfony\Component\Messenger\Transport\Serialization\PhpSerializer;
 
@@ -25,5 +26,29 @@ class PhpSerializerTest extends TestCase
         $envelope = new Envelope(new DummyMessage('Hello'));
 
         $this->assertEquals($envelope, $serializer->decode($serializer->encode($envelope)));
+    }
+
+    public function testDecodingFailsWithBadFormat()
+    {
+        $this->expectException(MessageDecodingFailedException::class);
+        $this->expectExceptionMessageRegExp('/Could not decode/');
+
+        $serializer = new PhpSerializer();
+
+        $serializer->decode([
+            'body' => '{"message": "bar"}',
+        ]);
+    }
+
+    public function testDecodingFailsWithBadClass()
+    {
+        $this->expectException(MessageDecodingFailedException::class);
+        $this->expectExceptionMessageRegExp('/class "ReceivedSt0mp" not found/');
+
+        $serializer = new PhpSerializer();
+
+        $serializer->decode([
+            'body' => 'O:13:"ReceivedSt0mp":0:{}',
+        ]);
     }
 }