+++ src/Symfony/Component/Messenger/Transport/AmqpExt/AmqpTransport.php
@@ -53,6 +53,22 @@ class AmqpTransport implements TransportInterface
     /**
      * {@inheritdoc}
      */
+    public function ack(Envelope $envelope): void
+    {
+        ($this->receiver ?? $this->getReceiver())->ack($envelope);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function reject(Envelope $envelope): void
+    {
+        ($this->receiver ?? $this->getReceiver())->reject($envelope);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function send(Envelope $envelope): Envelope
     {
         return ($this->sender ?? $this->getSender())->send($envelope);