+++ src/Symfony/Component/Messenger/Transport/AmqpExt/Connection.php
@@ -33,7 +33,7 @@ class Connection
         'x-message-ttl',
     ];
 
-    private $connectionCredentials;
+    private $connectionConfiguration;
     private $exchangeConfiguration;
     private $queueConfiguration;
     private $amqpFactory;
@@ -53,9 +53,47 @@ class Connection
      */
     private $amqpQueue;
 
-    public function __construct(array $connectionCredentials, array $exchangeConfiguration, array $queueConfiguration, AmqpFactory $amqpFactory = null)
+    /**
+     * @var \AMQPExchange|null
+     */
+    private $amqpDelayExchange;
+
+    /**
+     * Constructor.
+     *
+     * Available options:
+     *
+     *   * host: Hostname of the AMQP service
+     *   * port: Port of the AMQP service
+     *   * vhost: Virtual Host to use with the AMQP service
+     *   * user: Username to use to connect the the AMQP service
+     *   * password: Password to use the connect to the AMQP service
+     *   * queue:
+     *     * name: Name of the queue
+     *     * routing_key: The routing key (if any) to use to push the messages to
+     *     * flags: Queue flags (Default: AMQP_DURABLE)
+     *     * arguments: Extra arguments
+     *   * exchange:
+     *     * name: Name of the exchange
+     *     * type: Type of exchange (Default: fanout)
+     *     * flags: Exchange flags (Default: AMQP_DURABLE)
+     *     * arguments: Extra arguments
+     *   * delay:
+     *     * routing_key_pattern: The pattern of the routing key (Default: "delay_%delay%")
+     *     * queue_name_pattern: Pattern to use to create the queues (Default: "delay_queue_%delay%")
+     *     * exchange_name: Name of the exchange to be used for the retried messages (Default: "retry")
+     *   * auto-setup: Enable or not the auto-setup of queues and exchanges (Default: true)
+     *   * loop_sleep: Amount of micro-seconds to wait if no message are available (Default: 200000)
+     */
+    public function __construct(array $connectionConfiguration, array $exchangeConfiguration, array $queueConfiguration, AmqpFactory $amqpFactory = null)
     {
-        $this->connectionCredentials = $connectionCredentials;
+        $this->connectionConfiguration = array_replace_recursive([
+            'delay' => [
+                'routing_key_pattern' => 'delay_%delay%',
+                'exchange_name' => 'delay',
+                'queue_name_pattern' => 'delay_queue_%delay%',
+            ],
+        ], $connectionConfiguration);
         $this->exchangeConfiguration = $exchangeConfiguration;
         $this->queueConfiguration = $queueConfiguration;
         $this->amqpFactory = $amqpFactory ?: new AmqpFactory();
@@ -123,18 +161,100 @@ class Connection
     }
 
     /**
+     * @param int $delay The delay in milliseconds
+     *
      * @throws \AMQPException
      */
-    public function publish(string $body, array $headers = []): void
+    public function publish(string $body, array $headers = [], int $delay = 0): void
     {
+        if (0 !== $delay) {
+            $this->publishWithDelay($body, $headers, $delay);
+
+            return;
+        }
+
         if ($this->shouldSetup()) {
             $this->setup();
         }
 
         $flags = $this->queueConfiguration['flags'] ?? AMQP_NOPARAM;
-        $attributes = array_merge_recursive($this->queueConfiguration['attributes'] ?? [], ['headers' => $headers]);
+        $attributes = $this->getAttributes($headers);
 
         $this->exchange()->publish($body, $this->queueConfiguration['routing_key'] ?? null, $flags, $attributes);
+    }
+
+    /**
+     * @throws \AMQPException
+     */
+    private function publishWithDelay(string $body, array $headers = [], int $delay)
+    {
+        if ($this->shouldSetup()) {
+            $this->setupDelay($delay);
+        }
+
+        $routingKey = $this->getRoutingKeyForDelay($delay);
+        $flags = $this->queueConfiguration['flags'] ?? AMQP_NOPARAM;
+        $attributes = $this->getAttributes($headers);
+
+        $this->getDelayExchange()->publish($body, $routingKey, $flags, $attributes);
+    }
+
+    private function setupDelay(int $delay)
+    {
+        if (!$this->channel()->isConnected()) {
+            $this->clear();
+        }
+
+        $exchange = $this->getDelayExchange();
+        $exchange->declareExchange();
+
+        $queue = $this->createDelayQueue($delay);
+        $queue->declareQueue();
+        $queue->bind($exchange->getName(), $this->getRoutingKeyForDelay($delay));
+    }
+
+    private function getDelayExchange(): \AMQPExchange
+    {
+        if (null === $this->amqpDelayExchange) {
+            $this->amqpDelayExchange = $this->amqpFactory->createExchange($this->channel());
+            $this->amqpDelayExchange->setName($this->connectionConfiguration['delay']['exchange_name']);
+            $this->amqpDelayExchange->setType(AMQP_EX_TYPE_DIRECT);
+        }
+
+        return $this->amqpDelayExchange;
+    }
+
+    /**
+     * Creates a delay queue that will delay for a certain amount of time.
+     *
+     * This works by setting message TTL for the delay and pointing
+     * the dead letter exchange to the original exchange. The result
+     * is that after the TTL, the message is sent to the dead-letter-exchange,
+     * which is the original exchange, resulting on it being put back into
+     * the original queue.
+     */
+    private function createDelayQueue(int $delay)
+    {
+        $delayConfiguration = $this->connectionConfiguration['delay'];
+
+        $queue = $this->amqpFactory->createQueue($this->channel());
+        $queue->setName(str_replace('%delay%', $delay, $delayConfiguration['queue_name_pattern']));
+        $queue->setArguments([
+            'x-message-ttl' => $delay,
+            'x-dead-letter-exchange' => $this->exchange()->getName(),
+        ]);
+
+        if (isset($this->queueConfiguration['routing_key'])) {
+            // after being released from to DLX, this routing key will be used
+            $queue->setArgument('x-dead-letter-routing-key', $this->queueConfiguration['routing_key']);
+        }
+
+        return $queue;
+    }
+
+    private function getRoutingKeyForDelay(int $delay): string
+    {
+        return str_replace('%delay%', $delay, $this->connectionConfiguration['delay']['routing_key_pattern']);
     }
 
     /**
@@ -171,11 +291,6 @@ class Connection
         return $this->queue()->ack($message->getDeliveryTag());
     }
 
-    public function reject(\AMQPEnvelope $message): bool
-    {
-        return $this->queue()->reject($message->getDeliveryTag());
-    }
-
     public function nack(\AMQPEnvelope $message, int $flags = AMQP_NOPARAM): bool
     {
         return $this->queue()->nack($message->getDeliveryTag(), $flags);
@@ -196,8 +311,8 @@ class Connection
     public function channel(): \AMQPChannel
     {
         if (null === $this->amqpChannel) {
-            $connection = $this->amqpFactory->createConnection($this->connectionCredentials);
-            $connectMethod = 'true' === ($this->connectionCredentials['persistent'] ?? 'false') ? 'pconnect' : 'connect';
+            $connection = $this->amqpFactory->createConnection($this->connectionConfiguration);
+            $connectMethod = 'true' === ($this->connectionConfiguration['persistent'] ?? 'false') ? 'pconnect' : 'connect';
 
             try {
                 $connection->{$connectMethod}();
@@ -244,9 +359,9 @@ class Connection
         return $this->amqpExchange;
     }
 
-    public function getConnectionCredentials(): array
+    public function getConnectionConfiguration(): array
     {
-        return $this->connectionCredentials;
+        return $this->connectionConfiguration;
     }
 
     private function clear(): void
@@ -258,14 +373,19 @@ class Connection
 
     private function shouldSetup(): bool
     {
-        if (!\array_key_exists('auto-setup', $this->connectionCredentials)) {
+        if (!\array_key_exists('auto-setup', $this->connectionConfiguration)) {
             return true;
         }
 
-        if (\in_array($this->connectionCredentials['auto-setup'], [false, 'false'], true)) {
+        if (\in_array($this->connectionConfiguration['auto-setup'], [false, 'false'], true)) {
             return false;
         }
 
         return true;
+    }
+
+    private function getAttributes(array $headers): array
+    {
+        return array_merge_recursive($this->queueConfiguration['attributes'] ?? [], ['headers' => $headers]);
     }
 }