+++ src/Symfony/Component/Messenger/Transport/Receiver/ReceiverInterface.php
@@ -11,8 +11,12 @@
 
 namespace Symfony\Component\Messenger\Transport\Receiver;
 
+use Symfony\Component\Messenger\Envelope;
+use Symfony\Component\Messenger\Exception\TransportException;
+
 /**
  * @author Samuel Roze <samuel.roze@gmail.com>
+ * @author Ryan Weaver <ryan@symfonycasts.com>
  *
  * @experimental in 4.2
  */
@@ -23,6 +27,12 @@ interface ReceiverInterface
      *
      * The handler will have, as argument, the received {@link \Symfony\Component\Messenger\Envelope} containing the message.
      * Note that this envelope can be `null` if the timeout to receive something has expired.
+     *
+     * If the received message cannot be decoded, the message should not
+     * be retried again (e.g. if there's a queue, it should be removed)
+     * and a MessageDecodingFailedException should be thrown.
+     *
+     * @throws TransportException If there is an issue communicating with the transport
      */
     public function receive(callable $handler): void;
 
@@ -30,4 +40,18 @@ interface ReceiverInterface
      * Stop receiving some messages.
      */
     public function stop(): void;
+
+    /**
+     * Acknowledge that the passed message was handled.
+     *
+     * @throws TransportException If there is an issue communicating with the transport
+     */
+    public function ack(Envelope $envelope): void;
+
+    /**
+     * Called when handling the message failed and it should not be retried.
+     *
+     * @throws TransportException If there is an issue communicating with the transport
+     */
+    public function reject(Envelope $envelope): void;
 }