+++ src/Symfony/Component/Messenger/Transport/Receiver/StopWhenMessageCountIsExceededReceiver.php
@@ -52,4 +52,14 @@ class StopWhenMessageCountIsExceededReceiver implements ReceiverInterface
     {
         $this->decoratedReceiver->stop();
     }
+
+    public function ack(Envelope $envelope): void
+    {
+        $this->decoratedReceiver->ack($envelope);
+    }
+
+    public function reject(Envelope $envelope): void
+    {
+        $this->decoratedReceiver->reject($envelope);
+    }
 }