+++ src/Symfony/Component/Messenger/Transport/Receiver/StopWhenTimeLimitIsReachedReceiver.php
@@ -53,4 +53,14 @@ class StopWhenTimeLimitIsReachedReceiver implements ReceiverInterface
     {
         $this->decoratedReceiver->stop();
     }
+
+    public function ack(Envelope $envelope): void
+    {
+        $this->decoratedReceiver->ack($envelope);
+    }
+
+    public function reject(Envelope $envelope): void
+    {
+        $this->decoratedReceiver->reject($envelope);
+    }
 }