+++ src/Symfony/Component/Messenger/Transport/Serialization/PhpSerializer.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Messenger\Transport\Serialization;
 
 use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Exception\InvalidArgumentException;
+use Symfony\Component\Messenger\Exception\MessageDecodingFailedException;
 
 /**
  * @author Ryan Weaver<ryan@symfonycasts.com>
@@ -30,7 +31,7 @@ class PhpSerializer implements SerializerInterface
             throw new InvalidArgumentException('Encoded envelope should have at least a "body".');
         }
 
-        return unserialize($encodedEnvelope['body']);
+        return $this->safelyUnserialize($encodedEnvelope['body']);
     }
 
     /**
@@ -41,5 +42,36 @@ class PhpSerializer implements SerializerInterface
         return [
             'body' => serialize($envelope),
         ];
+    }
+
+    private function safelyUnserialize($contents)
+    {
+        $e = null;
+        $signalingException = new MessageDecodingFailedException(sprintf('Could not decode message using PHP serialization: %s.', $contents));
+        $prevUnserializeHandler = ini_set('unserialize_callback_func', self::class.'::handleUnserializeCallback');
+        $prevErrorHandler = set_error_handler(function ($type, $msg, $file, $line, $context = []) use (&$prevErrorHandler, $signalingException) {
+            if (__FILE__ === $file) {
+                throw $signalingException;
+            }
+
+            return $prevErrorHandler ? $prevErrorHandler($type, $msg, $file, $line, $context) : false;
+        });
+
+        try {
+            $meta = unserialize($contents);
+        } finally {
+            restore_error_handler();
+            ini_set('unserialize_callback_func', $prevUnserializeHandler);
+        }
+
+        return $meta;
+    }
+
+    /**
+     * @internal
+     */
+    public static function handleUnserializeCallback($class)
+    {
+        throw new MessageDecodingFailedException(sprintf('Message class "%s" not found during decoding.', $class));
     }
 }